Performance Fix

+++ src/logic/Component.java
@@ -3,7 +3,7 @@ package logic;
 import sharedObject.IRenderable;
 
 public abstract class Component implements IRenderable {
-    protected double x,y;
+    protected int x,y;
     protected int z;
     protected boolean visible;
     protected Component(){
@@ -12,6 +12,7 @@ public abstract class Component implements IRenderable {
     public int getZ(){
         return z;
     }
+
     @Override
     public boolean isVisible(){
         return visible;
